Changed do_update to take a float.
Also removed some dead code.

+++ src/system.cpp
@@ -3,7 +3,7 @@
 
 ISystem::ISystem(World& w) : world_(&w) {}
 
-void ISystem::update(int deltat) {
+void ISystem::update(float deltat) {
     do_update(deltat);
 }
